third_party/WebKit/LayoutTests/bluetooth/descriptor/readValue/read-succeeds.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('DisconnectingHealthThermometerAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => characteristic.getDescriptor(user_description.name))
    .then(descriptor => descriptor.readValue())
    .then(value => {
      let decoder = new TextDecoder('utf-8');
      let value_str = decoder.decode(value);
      assert_equals(value_str, "gatt.characteristic_user_description");
    });
}, 'A read request succeeds and returns the descriptor\'s value.');
</script>
